Declassify text wrapping
